Project refactor (#17)
* refactor

* refactor
