+++ package.json
@@ -11,15 +11,9 @@
     "@mui/icons-material": "^5.3.0",
     "@mui/material": "^5.2.8",
     "@mui/styles": "^5.3.0",
-    "@reduxjs/toolkit": "^1.7.1",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "@types/jest": "^27.4.0",
-    "@types/node": "^17.0.8",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-router-dom": "^5.3.2",
     "firebase": "^9.6.3",
     "prettier": "^2.5.1",
     "react": "^17.0.2",
@@ -29,7 +23,6 @@
     "react-scripts": "5.0.0",
     "sass": "^1.48.0",
     "start": "^5.1.0",
-    "typescript": "^4.5.4",
     "web-vitals": "^2.1.3",
     "yarn": "^1.22.17"
   },
