+++ src/App.jsx
@@ -1,0 +1,11 @@
+import { HomePage } from './pages';
+
+const App = () => {
+  return (
+    <>
+      <HomePage />
+    </>
+  );
+};
+
+export default App;
