--- src/components/Home.tsx
@@ -1,12 +1,0 @@
-import app from 'config/firebase';
-
- const Home = () => {
-  return (
-    <>
-      <h1>Home</h1>
-      <button onClick={() => app.auth().signOut()}></button>
-    </>
-  );
-};
-
-export default Home;
