+++ src/components/Layout/Layout.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import { NavigationBar } from '../NavigationBar';
+
+export const Layout = ({ children }) => {
+  return (
+    <>
+      <NavigationBar />
+      {children}
+    </>
+  );
+};
