--- src/components/Login.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Login = () => {
-  return <div></div>;
-};
-
-export default Login;
