+++ src/components/NavigationBar/NavigationBarStyle.js
