+++ src/components/index.js
@@ -1,0 +1,3 @@
+export * from './NavigationBar'
+export * from './Signup'
+export * from "./Layout";
