+++ src/index.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.scss';
 import App from './App';
 import { AuthProvider } from './provider/AuthProvider';
+import './styles/global.scss';
+;
+
 
 ReactDOM.render(
   <React.StrictMode>
