--- src/pages/HomePage.tsx
@@ -1,8 +1,0 @@
-const HomePage = () => {
-    return (
-    <div className="HomePage">
-    </div>
-    );
-};
-  
-export default HomePage;
