+++ src/provider/AuthProvider.jsx
@@ -1,0 +1,18 @@
+import { useEffect, useState } from 'react';
+
+import { AuthContext } from '../context/AuthContext';
+import { auth } from '../config/firebase';
+
+export const AuthProvider = ({ children }) => {
+  const [user, setUser] = useState();
+
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((firebaseUser) => {
+      setUser(firebaseUser);
+    });
+
+    return unsubscribe;
+  }, []);
+
+  return <AuthContext.Provider value={user}>{children}</AuthContext.Provider>;
+};
