--- src/react-app-env.d.ts
@@ -1,14 +1,0 @@
-/// <reference types="react-scripts" />
-declare module '*.png';
-declare module '*.scss';
-declare module '*.svg';
-declare module '@rebass/preset';
-declare module '@rebass/forms';
-
-// declare namespace NodeJS {
-//   interface ProcessEnv {
-//     NODE_ENV: "development" | "production" | "test";
-//     PUBLIC_URL: string;
-//     REACT_APP_API_SERVER: string;
-//   }
-// }
