--- src/state/Puppylog/puppylogSlice.ts
@@ -1,9 +1,0 @@
-import { createSlice } from "@reduxjs/toolkit";
-
-export const slice = createSlice({
-  name: "puppylog",
-  initialState: {},
-  reducers: {},
-});
-
-export default slice.reducer;
