--- src/state/store.ts
@@ -1,8 +1,0 @@
-import { configureStore } from '@reduxjs/toolkit';
-import puppylogReducer from './Puppylog/puppylogSlice';
-
-export default configureStore({
-  reducer: {
-    puppylog: puppylogReducer,
-  },
-})
