+++ src/styles/global.scss
@@ -1,3 +1,8 @@
-@import "./variables";
-@import "./normalize";
-@import "./fonts";
+@import './variables';
+@import './normalize';
+@import './fonts';
+
+body {
+  padding: 0;
+  margin: 0;
+}
